+++ changelog.md
@@ -1,9 +1,12 @@
 == Changelog ==
 
-= Version 2.6.2 – September 5, 2024 =
+= Version 2.6.2 – September 7, 2024 =
 
++   Purge expired transients on cache clearing and automatically (daily).
 +   New `text_to_array()` function to split textarea to array of lines.
-+   Enhance security extension...
++   New AbuseIPDB api (security) extension to block by IP address based on abuse score.
+    +   See : https://www.abuseipdb.com 
++   Enhanced security extension...
     +   Block REST index list, WP core REST routes, non-rest json requests.
     +   CORS headers w/white-list domains.
 +   Updated wpconfig-transformer to v1.3.6